我的直播列表添加返回字段

+++ live-api/src/main/java/com/me2me/live/service/LiveService.java
@@ -36,14 +36,14 @@ public interface LiveService {
      * 获取直播列表
      * @return
      */
-    Response getMyLives(long uid,int sinceId);
+    Response getMyLives(long uid,long sinceId);
 
     /**
      * 获取所有未结束的直播列表
      * @param uid
      * @return
      */
-    Response getLives(long uid,int sinceId);
+    Response getLives(long uid,long sinceId);
 
     /**
      * 关注/取消关注我参与的直播
