+++ root/src/main/java/com/me2me/web/request/GetLivesRequest.java
@@ -14,5 +14,5 @@ public class GetLivesRequest extends Request{
 
     @Getter
     @Setter
-    private int sinceId;
+    private long sinceId;
 }
