+++ root/src/main/java/com/me2me/web/request/GetMyLivesRequest.java
@@ -14,5 +14,5 @@ public class GetMyLivesRequest extends Request{
 
     @Setter
     @Getter
-    private int sinceId;
+    private long sinceId;
 }
